Move query outside of table component
