+++ static/app/views/starfish/views/webServiceView/failureDetailPanel/index.tsx
@@ -1,4 +1,9 @@
+import {DEFAULT_STATS_PERIOD} from 'sentry/constants';
 import {t} from 'sentry/locale';
+import {NewQuery} from 'sentry/types';
+import DiscoverQuery from 'sentry/utils/discover/discoverQuery';
+import EventView from 'sentry/utils/discover/eventView';
+import {decodeScalar} from 'sentry/utils/queryString';
 import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
 import Detail from 'sentry/views/starfish/components/detailPanel';
@@ -15,6 +20,32 @@ export default function FailureDetailPanel({
   const location = useLocation();
   const organization = useOrganization();
 
+  const {query} = location;
+  const hasStartAndEnd = query.start && query.end;
+  const newQuery: NewQuery = {
+    name: t('Failure Sample'),
+    projects: [],
+    start: decodeScalar(query.start),
+    end: decodeScalar(query.end),
+    range: !hasStartAndEnd
+      ? decodeScalar(query.statsPeriod) || DEFAULT_STATS_PERIOD
+      : undefined,
+    fields: [
+      'transaction',
+      'count_if(http.status_code,greaterOrEquals,500)',
+      'equation|count_if(http.status_code,greaterOrEquals,500)/(count_if(http.status_code,equals,200)+count_if(http.status_code,greaterOrEquals,500))',
+      'http.method',
+      'count_if(http.status_code,equals,200)',
+    ],
+    query:
+      'event.type:transaction has:http.method transaction.op:http.server count_if(http.status_code,greaterOrEquals,500):>0',
+    version: 2,
+  };
+
+  newQuery.orderby = '-count_if_http_status_code_greaterOrEquals_500';
+
+  const eventView = EventView.fromNewQueryWithLocation(newQuery, location);
+
   return (
     <Detail detailKey={spike?.startTimestamp.toString()} onClose={onClose}>
       <div>
@@ -24,7 +55,27 @@ export default function FailureDetailPanel({
             'Detailed summary of failure rate spike. Detailed summary of failure rate spike. Detailed summary of failure rate spike. Detailed summary of failure rate spike. Detailed summary of failure rate spike. Detailed summary of failure rate spike.'
           )}
         </p>
-        <FailureDetailTable location={location} organization={organization} />
+        {spike && (
+          <DiscoverQuery
+            eventView={eventView}
+            orgSlug={organization.slug}
+            location={location}
+            referrer="api.starfish.failure-event-list"
+            queryExtras={{dataset: 'discover'}}
+            limit={5}
+          >
+            {results => {
+              return (
+                <FailureDetailTable
+                  {...results}
+                  location={location}
+                  organization={organization}
+                  eventView={eventView}
+                />
+              );
+            }}
+          </DiscoverQuery>
+        )}
       </div>
     </Detail>
   );
